Remove ProductFavourite
