+++ src/components/common/ProductBox/ProductBoxContainer.js
@@ -1,6 +1,7 @@
 import { connect } from 'react-redux';
 import ProductBox from './ProductBox';
 import { setCompare, getCount } from '../../../redux/compareRedux';
+import { addToFavourite } from '../../../redux/productsRedux';
 
 const mapStateToProps = state => ({
   count: getCount(state),
@@ -9,6 +10,7 @@ const mapStateToProps = state => ({
 
 const mapDispatchToProps = dispatch => ({
   setCompare: value => dispatch(setCompare(value)),
+  addToFavourite: value => dispatch(addToFavourite(value)),
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(ProductBox);
