--- src/components/features/ProductFavourite/ProductFavourite.module.scss
@@ -1,6 +1,0 @@
-@import "../../../styles/settings.scss";
-
-.heart {
-  color: $favourite-color;
-  background-color: $primary;
-}
