+++ src/redux/productsRedux.js
@@ -14,11 +14,11 @@ const createActionNameFav = name => `products/heart/${name}`;
 
 /* action types */
 export const SET_FAVOURITE = createActionNameFav('SET_FAVOURITE');
-export const UNSET_FAVOURITE = createActionNameFav('UNSET_FAVOURITE');
+// export const UNSET_FAVOURITE = createActionNameFav('UNSET_FAVOURITE');
 
 /* action creator */
 export const addToFavourite = payload => ({ payload, type: SET_FAVOURITE });
-export const removeFromFavourite = payload => ({ payload, type: UNSET_FAVOURITE });
+// export const removeFromFavourite = payload => ({ payload, type: UNSET_FAVOURITE });
 
 /*SECTION STARS*/
 
@@ -49,7 +49,8 @@ export default function reducer(statePart = [], action = {}) {
     case SET_FAVOURITE: {
       const newStatePart = statePart.map(product => {
         if (product.id === action.payload.id) {
-          product.heart = true;
+          product.heart = !product.heart;
+          // product.heart = true;
           return product;
         } else {
           return product;
@@ -57,17 +58,17 @@ export default function reducer(statePart = [], action = {}) {
       });
       return newStatePart;
     }
-    case UNSET_FAVOURITE: {
-      const newStatePart = statePart.map(product => {
-        if (product.id === action.payload.id) {
-          product.heart = false;
-          return product;
-        } else {
-          return product;
-        }
-      });
-      return newStatePart;
-    }
+    // case UNSET_FAVOURITE: {
+    //   const newStatePart = statePart.map(product => {
+    //     if (product.id === action.payload.id) {
+    //       product.heart = false;
+    //       return product;
+    //     } else {
+    //       return product;
+    //     }
+    //   });
+    //   return newStatePart;
+    // }
     default:
       return statePart;
   }
